Moving Types to a seperate folder
