+++ src/Components/Names/Gender.hpp
@@ -1,12 +1,8 @@
 #pragma once
-#include <cstdint>
+
+#include <Types/Gender.hpp>
 
 namespace pokesim {
-namespace dex {
-/** @brief Pokemon gender name */
-enum Gender : std::uint8_t { NO_GENDER = 0, FEMALE, MALE, GENDER_TOTAL };
-}  // namespace dex
-
 struct GenderName {
   dex::Gender name = dex::NO_GENDER;
 };
