+++ src/Components/Names/Status.hpp
@@ -1,12 +1,8 @@
 #pragma once
-#include <cstdint>
+
+#include <Types/Status.hpp>
 
 namespace pokesim {
-namespace dex {
-/** @brief Pokemon status condition name */
-enum Status : std::uint8_t { NO_STATUS = 0, BRN, FRZ, PAR, PSN, SLP, TOX, /*, FRB, DRO, */ STATUS_TOTAL };
-}  // namespace dex
-
 struct StatusName {
   dex::Status name = dex::NO_STATUS;
 };
