+++ src/Components/Names/Type.hpp
@@ -1,18 +1,8 @@
 #pragma once
-#include <cstdint>
+
+#include <Types/Type.hpp>
 
 namespace pokesim {
-namespace dex {
-// clang-format off
-
-/** @brief Pokemon type name */
-enum Type : std::uint8_t {
-  NO_TYPE = 0, NORMAL_TYPE, FIGHTING_TYPE, FLYING_TYPE, POISON_TYPE, GROUND_TYPE, ROCK_TYPE, BUG_TYPE, GHOST_TYPE, STEEL_TYPE, FIRE_TYPE, WATER_TYPE, GRASS_TYPE, ELECTRIC_TYPE, PSYCHIC_TYPE, ICE_TYPE, DRAGON_TYPE, DARK_TYPE, FAIRY_TYPE, TYPE_TOTAL
-};
-
-// clang-format on
-}  // namespace dex
-
 struct TypeName {
   dex::Type name = dex::NO_TYPE;
 };
