+++ src/Dex/SetupClasses/ItemSetup.cpp
@@ -1,5 +1,7 @@
 #include "ItemSetup.hpp"
 
+#include <Components/Names/Item.hpp>
+
 namespace pokesim::dex::internal {
 void ItemSetup::setName(Item item) {
   handle.emplace<ItemName>(item);
