+++ src/Dex/SetupClasses/MoveSetup.cpp
@@ -6,6 +6,8 @@
 #include <Components/Chance.hpp>
 #include <Components/EntityHolders/MoveEffect.hpp>
 #include <Components/MultiHit.hpp>
+#include <Components/Names/Move.hpp>
+#include <Components/Names/Type.hpp>
 #include <Components/PP.hpp>
 #include <Components/Priority.hpp>
 #include <Components/Tags/Move.hpp>
