+++ src/PokeSim.hpp
@@ -3,3 +3,4 @@
 #include <Components/Components.hpp>
 #include <Dex/Dex.hpp>
 #include <Dex/Names.hpp>
+#include <Types/Types.hpp>
