+++ src/Types/Mechanics.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <cstdint>
+
+namespace pokesim {
+enum Mechanics : std::uint32_t {
+  SCARLET_VIOLET,
+};
+}
