+++ src/Types/Types.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "Ability.hpp"
+#include "Gender.hpp"
+#include "Item.hpp"
+#include "Mechanics.hpp"
+#include "Move.hpp"
+#include "Nature.hpp"
+#include "Species.hpp"
+#include "Stat.hpp"
+#include "Status.hpp"
+#include "Type.hpp"
